refactor: extract PDF prose boundary evidence
